+++ plugin/app/Http/Controllers/Api/LabController.php
@@ -6,9 +6,11 @@ use Carbon\Carbon;
 use Illuminate\Http\Request;
 use TTU\Charon\Http\Controllers\Controller;
 use TTU\Charon\Models\Charon;
+use TTU\Charon\Models\CharonDefenseLab;
 use TTU\Charon\Models\Lab;
 use TTU\Charon\Repositories\LabRepository;
 use TTU\Charon\Services\LabService;
+use Zeizig\Moodle\Globals\User;
 use Zeizig\Moodle\Models\Course;
 
 class LabController extends Controller
@@ -91,7 +93,7 @@ class LabController extends Controller
             $this->request['name'],
             $this->request['teachers'],
             $this->request['charons'],
-            $this->request['groups'],
+            $this->request['groups']
         );
     }
 
@@ -159,9 +161,13 @@ class LabController extends Controller
         return ['groups' => $groups, 'groupings' => $result];
     }
 
-    public function findLabsByCharonLaterEqualToday(Request $request)
+    /** Function what will return list of defense labs with lists of students - defenders registered for each lab
+     * @param Request $request
+     * @return mixed
+     */
+    public function findUpcomingOrActiveLabsByCharon(Charon $charon)
     {
-        return $this->labService->findLabsByCharonIdLaterEqualToday($request->route('charon'));
+        return $this->labService->findUpcomingOrActiveLabsByCharon($charon->id);
     }
 
     /**
@@ -179,4 +185,15 @@ class LabController extends Controller
         return $this->labRepository->countRegistrations($lab->id, $start, $end, $charons, $teachers);
     }
 
+    /**
+     * Returns queue status in the form of an array, with approximate defence times.
+     *
+     * @param Charon $charon
+     * @param CharonDefenseLab $defenseLab
+     * @return array
+     */
+    public function getLabQueueStatus(Charon $charon, CharonDefenseLab $defenseLab)
+    {
+        return $this->labService->labQueueStatus(app(User::class)->currentUser(), $defenseLab->lab);
+    }
 }