+++ plugin/app/Repositories/DefenseRegistrationRepository.php
@@ -307,4 +307,31 @@ class DefenseRegistrationRepository
             ->get();
     }
 
+    /**
+     * @param int $labId
+     * @return mixed
+     */
+    public function getListOfLabRegistrationsByLabId(int $labId)
+    {
+        return DB::table('charon_defenders')
+            ->join("charon", "charon.id", "charon_defenders.charon_id")
+            ->join("charon_defense_lab","charon_defense_lab.id","charon_defenders.defense_lab_id")
+            ->where("charon_defense_lab.lab_id", $labId)
+            ->select("charon.name as charon_name", "charon.defense_duration as charon_length", "charon_defenders.student_id")
+            ->orderBy("charon_defenders.id")
+            ->get()
+            ->all();
+    }
+
+    /**
+     * @param int $labId
+     * @return mixed
+     */
+    public function countDefendersByLab(int $labId)
+    {
+        return DB::table('charon_defenders')
+            ->where ('defense_lab_id', $labId) //where id of defense lab equals to id of lab sending by function
+            ->count();
+    }
+
 }