+++ plugin/routes/api.php
@@ -93,7 +93,8 @@ Route::group(['namespace' => 'Api'], function () {
 
     // LABS
 
-    Route::get('charons/{charon}/labs/view', 'LabController@findLabsByCharonLaterEqualToday'); // get labs student can register to
+    Route::get('charons/{charon}/labs/available', 'LabController@findUpcomingOrActiveLabsByCharon');
+    // get labs which are available for registration
     Route::middleware('auth.course.managing.require')  // save lab
         ->post('courses/{course}/labs', 'LabController@save');
     Route::middleware('auth.course.managing.require')  // get all labs for course
@@ -104,9 +105,11 @@ Route::group(['namespace' => 'Api'], function () {
         ->delete('courses/{course}/labs/{lab}', 'LabController@delete');
     Route::middleware('auth.course.managing.require')  // update lab
         ->post('courses/{course}/labs/{lab}/update', 'LabController@update');
+    // get number of affected registrations when lab is being to deleted or modified
     Route::middleware('auth.course.managing.require')
-        ->get('courses/{course}/labs/{lab}/registrations', 'LabController@countRegistrations'); 
-        // get number of affected registrations when lab is being to deleted or modified
+        ->get('courses/{course}/labs/{lab}/registrations', 'LabController@countRegistrations');
+    Route::middleware('auth.charon.submissions.view.require') // get lab queue status
+        ->get('charons/{charon}/defenseLab/{defenseLab}/queueStatus', 'LabController@getLabQueueStatus');
 
     // TEACHERS